data comes back from strava API, display bugs sorted out

+++ app/src/main/java/com/example/rideyourbike/data/remote/dto/ActivitiesDTOItem.kt
@@ -15,7 +15,7 @@ data class ActivitiesDTOItem(
     @SerializedName("athlete_count")
     val athleteCount: Int,
     @SerializedName("average_heartrate")
-    val averageHeartrate: Int,
+    val averageHeartrate: Double,
     @SerializedName("average_speed")
     val averageSpeed: Double,
     @SerializedName("comment_count")
