+++ app/src/main/java/com/example/rideyourbike/domain/model/ActivityDisplayItem.kt
@@ -4,7 +4,7 @@ data class ActivityDisplayItem(
     val name: String = "",
     val type: String = "",
     val distance: Double = 0.0,
-    val averageHeartRate: Int = 0,
+    val averageHeartRate: Double = 0.0,
     val maxHeartRate: Int = 0,
     val hasKudos: Boolean = false
 )
